+++ web/components/shared/metrics/barChart.tsx
@@ -11,8 +11,8 @@ import {
   YAxis,
 } from "recharts";
 import { ValueType } from "recharts/types/component/DefaultTooltipContent";
-import { getUSDate, getUSDateShort } from "../../shared/utils/utils";
-import { clsx } from "../../shared/clsx";
+import { getUSDate, getUSDateShort } from "../utils/utils";
+import { clsx } from "../clsx";
 
 export interface BarChartData {
   time: Date;
