+++ web/components/templates/dashboard/panels/costsPanel.tsx
@@ -1,7 +1,7 @@
 import { UserGroupIcon } from "@heroicons/react/24/outline";
 import { Result } from "../../../../lib/result";
 import { CostOverTime } from "../../../../pages/api/metrics/costOverTime";
-import { RenderBarChart } from "../barChart";
+import { RenderBarChart } from "../../../shared/metrics/barChart";
 import { Loading } from "../dashboardPage";
 
 interface CostPanelProps {
