+++ web/components/templates/dashboard/panels/errorsPanel.tsx
@@ -1,8 +1,8 @@
 import { UserGroupIcon } from "@heroicons/react/24/outline";
 import { useErrorPageOvertime } from "../../../../services/hooks/useCachePage";
 import { useErrorPageCodes } from "../../../../services/hooks/useErrorPage";
-import ThemedPieChart from "../../cache/modelPIeChart";
-import { MultilineRenderLineChart } from "../../cache/timeGraph";
+import { StackedBarChart } from "../../../shared/metrics/stackedBarChart";
+import { RenderPieChart } from "../../../shared/metrics/pieChart";
 
 interface ErrorsPanelProps {}
 
@@ -13,7 +13,28 @@ const ErrorsPanel = (props: ErrorsPanelProps) => {
 
   const errorCodesOverTime = useErrorPageOvertime();
 
-  console.log(errorCodesOverTime.overTime.data);
+  const data = errorCodesOverTime.overTime.data?.data ?? [];
+
+  const timeMap = (x: Date) => new Date(x).toDateString();
+
+  const chartData = data.map((d) => ({
+    ...d,
+    time: timeMap(d.time),
+  }));
+
+  const getErrorCodes = () => {
+    const errorCodes = new Set<string>();
+    chartData.forEach((d) => {
+      Object.keys(d).forEach((key) => {
+        if (key !== "time") {
+          errorCodes.add(key);
+        }
+      });
+    });
+    return Array.from(errorCodes);
+  };
+
+  const errorCodes = getErrorCodes();
 
   return (
     <div className="grid grid-cols-5 gap-4 h-96">
@@ -23,31 +44,38 @@ const ErrorsPanel = (props: ErrorsPanelProps) => {
             Errors
           </h3>
           <div className="h-72 px-4">
-            <MultilineRenderLineChart
-              data={errorCodesOverTime.overTime.data?.data ?? []}
-              timeMap={(x) => new Date(x).toDateString()}
-              valueFormatter={(x) => `${x}`}
-            />
+            {errorCodesOverTime.overTime.isLoading ? (
+              <div className="h-full w-full flex-col flex p-8">
+                <div className="h-full w-full rounded-lg bg-gray-300 animate-pulse" />
+              </div>
+            ) : (
+              <StackedBarChart data={chartData} keys={errorCodes} />
+            )}
           </div>
         </div>
       </div>
       <div className="col-span-2 bg-white border border-gray-300 rounded-lg">
         <div className="flex flex-col space-y-4 py-6">
           <h3 className="text-lg font-semibold text-gray-900 text-center">
-            Errors
+            Distribution
           </h3>
           <div className="h-72">
-            <ThemedPieChart
-              data={
-                pageCodes.errorCodes.data?.data
-                  ?.filter((x) => x.error_code !== 200)
-                  .map((x) => ({
-                    name: "" + x.error_code,
-                    value: +x.count,
-                  })) ?? []
-              }
-              isLoading={pageCodes.errorCodes.isLoading}
-            />
+            {pageCodes.errorCodes.isLoading ? (
+              <div className="h-full w-full flex-col flex p-8">
+                <div className="h-full w-full rounded-lg bg-gray-300 animate-pulse" />
+              </div>
+            ) : (
+              <RenderPieChart
+                data={
+                  pageCodes.errorCodes.data?.data
+                    ?.filter((x) => x.error_code !== 200)
+                    .map((x) => ({
+                      name: "" + x.error_code,
+                      value: +x.count,
+                    })) ?? []
+                }
+              />
+            )}
           </div>
         </div>
       </div>
