+++ web/pages/stats.tsx
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useRouter } from "next/router";
 import BasePageV2 from "../components/shared/layout/basePageV2";
 import MetaData from "../components/shared/metaData";
-import { RenderBarChart } from "../components/templates/dashboard/barChart";
+import { RenderBarChart } from "../components/shared/metrics/barChart";
 import { BaseUrlInstructions } from "../components/templates/welcome/welcomePage";
 import { Result } from "../lib/result";
 import { HeliconeStats } from "./api/stats";
